fixed tests according to changed controllers structure
