+++ backend/tests/controller.test.js
@@ -232,53 +232,3 @@ describe('user_survey_answers', () => {
     expect(surveyAnswers[0].name).toEqual('1a2y')
   })
 })
-
-jest.mock('../models/user_activities', () => () => {
-  const SequelizeMock = require('sequelize-mock')
-  const dbMock = new SequelizeMock()
-
-  return dbMock.define('user_activities', {
-    id: 87,
-    user_id: 'fdd8sklla'
-  })
-})
-
-
-describe('user_activities today', () => {
-  it('are returned correctly', async () => {
-    const activitiesToday = await controller.findUserActivitiesToday()
-    const createdAt = (activitiesToday[0].createdAt).setHours(0, 0, 0)
-    const timeNow = new Date().setHours(0, 0, 0, 0)
-    expect(activitiesToday.length).toEqual(1)
-    expect(createdAt - timeNow).toBeLessThan(1000)
-  })
-})
-
-jest.mock('../models/user_profiles', () => () => {
-  const SequelizeMock = require('sequelize-mock')
-  const dbMock = new SequelizeMock()
-
-  return dbMock.define('user_profiles', {
-    user_id: '1a2b3c',
-    height: '',
-    weight: '',
-    sex: null,
-    birth_date: null,
-    first_name: 'Mikko',
-    last_name: 'Mallikas',
-    added_organisation: 'OHTU'
-  })
-})
-
-describe('new users from the last seven days', () => {
-  it('are returned correctly', async () => {
-    const newUsers = await controller.findNewUsers()
-    const createdAt = (newUsers[0].createdAt).setHours(0, 0, 0)
-    const timeNow = new Date().setHours(0, 0, 0, 0)
-    expect(newUsers.length).toEqual(1)
-    expect(createdAt - timeNow).toBeLessThan(1000)
-  })
-})
-
-
-
